various changes and better memory handling
