+++ blacklist.txt
@@ -1,0 +1,3 @@
+# Blacklist for ASAN
+# ignore everything in third_party
+src:third_party/*
