+++ geometry.cpp
@@ -3,11 +3,7 @@
 //
 #include "geometry.h"
 
-template class Vec2<float>;
-template class Vec3<float>;
-template class Matrix<float>;
-template class Matrix4x4<float>;
-template class Matrix3x3<float>;
+
 
 /// Create a homogonized matrix from a vector
 Vec4f homogonize(Vec3f const &v, float h) {
@@ -134,3 +130,8 @@ Matrix<T> Matrix<T>::operator*(const Vec3<T> &other) {
     }
     return result;
 }
+template class Vec2<float>;
+template class Vec3<float>;
+template class Matrix<float>;
+template class Matrix4x4<float>;
+template class Matrix3x3<float>;
